+++ src/main/scala/org/podval/tools/test/filter/TestsTestFilterMatch.scala
@@ -2,16 +2,16 @@ package org.podval.tools.test.filter
 
 final case class TestsTestFilterMatch(
   testNames: Set[String],
-  testWildCards: Set[String]
+  testWildcards: Set[String]
 ) extends TestFilterMatch:
-  override def explicitlySpecified: Boolean = true
+  override def explicitlySpecified: Boolean = false
 
-  override def isEmpty: Boolean = testNames.isEmpty && testWildCards.isEmpty
+  override def isEmpty: Boolean = testNames.isEmpty && testWildcards.isEmpty
 
   override def intersect(other: TestFilterMatch): TestFilterMatch = other match
     case _   : SuiteTestFilterMatch => this
     case that: TestsTestFilterMatch => TestsTestFilterMatch(
       testNames = this.testNames.intersect(that.testNames) ++
-        this.testNames.filter(l => that.testWildCards.exists(l.contains)), // TODO [filter]
-      testWildCards = this.testWildCards.intersect(that.testWildCards)
+        this.testNames.filter(l => that.testWildcards.exists(l.contains)), // TODO [filter]
+      testWildcards = this.testWildcards.intersect(that.testWildcards)
     )