+++ src/main/scala/org/podval/tools/test/framework/ZioTest.scala
@@ -55,7 +55,7 @@ object ZioTest extends FrameworkDescriptor(
   displayName = "ZioTest",
   group = "dev.zio",
   artifact = "zio-test-sbt",
-  versionDefault = Version("2.1.16"),
+  versionDefault = Version("2.1.17"),
   className = "zio.test.sbt.ZTestFramework",
   sharedPackages = List("zio.test.sbt"),
   tagOptionStyle = OptionStyle.OptionPerValue,