+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cleanup: `ScalaJSPlugin` and `BackendDelegate`s;
 - cleanup: introduced `Running`;
 - cleanup: introduced `FrameworkProvider`;
+- cleanup: introduced `Runners`;
+- cleanup: introduced `SkippedTests`;
 - cleanup: `RunTestClassProcessor`;
 - dependency updates;
 - doc: `README` improvements;